Separate Job (interface) and JobSupport (implementation). No other changes except making JobImpl internal (not private)
